pyproject.toml
[project]
name = "jack_midi_monitor"
authors = [{name = "Leon Dionne", email = "ldionne@dridesign.sh.cn"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
	"License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3.10",
	"Topic :: Multimedia :: Sound/Audio :: MIDI",
	"Topic :: Utilities"
]
dynamic = ["version", "description"]
dependencies = ["JACK-Client", "midi_notes", "PyQt5", "soso_qt_extras >= 1.2.0"]

[project.scripts]
jack-midi-monitor-cli = "jack_midi_monitor.__main__:main"
jack-midi-monitor = "jack_midi_monitor.gui:main"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[bumpver]
current_version = "1.1.2"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag = false
push = false

[bumpver.file_patterns]
"jack_midi_monitor/__init__.py" = [
    '^__version__ = "{version}"$',
]
